+++ wezterm-client/src/client.rs
@@ -1361,6 +1361,11 @@ impl Client {
     );
     rpc!(get_lines, GetLines, GetLinesResponse);
     rpc!(
+        get_semantic_zones,
+        GetSemanticZones,
+        GetSemanticZonesResponse
+    );
+    rpc!(
         get_dimensions,
         GetPaneRenderableDimensions,
         GetPaneRenderableDimensionsResponse